Move OtherCtrls to end to make enum testing easier
On Windows using VS Code debugger, this generates an exception in Windows -- just as it does for any other common dialog that calls into Windows for file/folder dialogs. Since I want to have exceptions always turned on, we move this to the bottom of the list so that we can quickly step through every form to confirm Mockup is working correctly.