Change strings to enumerations for lookup (#136)
* Change namespace prop to name_space

We can't use "namespace" as an enumeration value without confusing the compiler

* Add warning about importance of updating enum_classes

Once we start switching over to the enumerated types, it will be critical to keep them current. Note that is was also critical to always use the correct spelling before, with the only difference being that you got no warning if something wasn't entered correctly.

* Eliminate old Type enumeration

* Major overhaul of NodeType class and most callers that use it.

* Complete moving NodeDeclarations into an array

Now allocated with new instead of std::shared_ptr, deleted in new NodeCreator dtor .

* Change "default" property to "default_btn "

The clang compiler will not allow the name "default" to be used as a enumeration value, even in a scoped enum. That may break wxFormBuilder import, but that needs to be a separate issue -- I don't want to complicate this PR any more than necessary with unrelated changes.

* Move NodeCreator initialization to a new file and optimize

Switched from strings to enumerations, and combine functions to prevent rereading the same XML files.

* Move wxUiTesting OtherCtrls dialog to end of project to make enum testing easier

On Windows using VS Code debugger, this generates an exception in Windows -- just as it does for any other common dialog that calls into Windows for file/folder dialogs. Since I want to have exceptions always turned on, we move this to the bottom of the list so that we can quickly step through every form to confirm Mockup is working correctly.

Co-authored-by: KeyWorksRW <42786259+KeyWorksRW@users.noreply.github.com>