Use hash value to identify internal/external clipboard data
Hash is generated from node name and all property values of the node and all children, then stored as the  wxUiEditor format in the clipboard. as well as in MainFrame. On Paste, this value is compared between the external and internal clipboard hashes, and if they match the internal clipboard is used to prevent unnecessary memory allocation and XML parsing.