Support inserting a column or row
There are now two hidden radio buttons that are only shown when the infobar is shown. The radio buttons determine which type of insertion to perform.

The close button in the infobar has been replaced with an Insert button that will close the dialog after setting the m_action member along with the standard members saved when OK is clicked.