Add flags to check if Undo/Redo generated events
Some of the derived undo command classes already fire the necessary events, so there's no need for the Undo() or Redo() callers to fire the same events. This also avoids unnecessary calls to FireProjectUpdatedEvent() which causes the entire tree control in the navigation panel to be recreated. Avoiding that case should help a lot with performance.