Add UI to control display of creation functions
The Navigation, Property, and Mockup panels all have a call to MSG_INFO() when the contents of the panel is created (or recreated). This is used to check that we aren't accidentally recreating the contents when we don't need to.

This commit adds UI to the Debug Settings dialog that controls whether or not to call MSG_INFO(). That makes it easy to test without having to recompile.