Change form type checks to a single function using iterator list
Previously, adding a new form type required changing code in multiple places. This changes that to use a single list of form types, and functions can iterate through that list. That way when a new form type is added, all you have to do is add the type to the list.