Tweak some ASSERTs to be more informative
Mostly this is changing ASSERT(parent) to ASSERT(node->GetParent()) -- this has the advantage of how the parent is being retrieved.

I also added some more code conditions in Mockup to prevent crashing in release. I can't find a reproducible case, but every once in a while a selection gets screwed up and still sent to Mockup even though the selected node is now parentless. Previously that would cause a crash, now it is silently ignored in release.