Default to using dialog units for -1,-1 point/size
Dialog units are the preferred way to set sizes and positions, so if a default value is in place, then set dialog units. It will be ignored as long as it remains a default value, but if the user changes one of the values, then the use dialog units will already be set for them.

Note that we can *not* change existing values to use dialog units -- that would break imports of wxFormBuilder and wxSmith that don't support dialog units.