Add code to convert option/bitlist to friendly name in old projects
This removes the old version 11 fixup, adds a new one that makes it possible to change any option or bitlist types and have them automatically converted to a new friendly name.

This is currently set for version 14, but only until we've updated any properties, then it should be dropped to <13.