Search for Bind if Event properties are shown
If the event panel in the property grid is show, first see if the node has a Bind event, and if so, mark and display it. Otherwise, display the node creation the way it normally would be displayed.

Switching between the Event and Properties page will switch the scintilla bookmark (if an event handler has been defined).