Convert form wxPanel to 2-step construction
Similar to wxDialog and most wxWidget classes, this allows either an empty constructor that then calls Create(...) or a full constructor which calls create in it's declaration.

Note that this is the only way to import a form panel from the current wxFormBuilder if the property two_step_creation is checked.