Remove temp files in prop_base_directory
If prop_base_directory has been set, then the temporary files will be created
in that directory instead of the project file directory. This commit will check
if the property has been specified, and remove the files there if it has.

This is a Debug only change, since it is part of the Compare Code files
command.