Convert svg files to zlib compressed embed data
This will strip out most unneeded nodes and attributes from the xml file, convert it to a single string with no line breaks, and then compress it into zlip data. The data is stored the same way we do with embedded images. Reading it back means decompressing the zlib data into a string and passing it off to wxBitmapBundle::FromSvg().