Pass ttlib::multistr parts instead of description
Image properties contain multiple parts separated with a ';'.
We call ttlib::multistr to break them apart to work with the
individual components. This commit enables pass the
parts class as a parameter instead of the full property
string. This is to avoid having to separate out the parts
multiple times.