Fix problems from testing
Note that we now write the wxueImage() function declaration
in the Image form generated header file. Conditionalized code
can use either the bundle loading call, or call this function instead.
Having it declared in the header file means we don't need to
define it in any source file that includes the Image form header
file.