Add option to build from wxWidgets fork
Realistically, I'm the only one that can use this since it requires
a fork at the same level, and a bld/ folder with a custom
CMakeLists.txt file. I added it so that I can confirm a build will
work with a version of wxWidgets that has not been released
yet.