Support embedded .xpm files
Don't use the extension to determine image type. It is perfectly
valid to Embed a .xpm file, which will automatically convert it to
a PNG image. This is actually the preferred method for using
XPM images since it avoids the c++20 error about a literal string
being assigned to a char array, and we don't need to know what
name was used inside the file. See issue #716