Fire node selected event when project loaded
This commit results in the Project node always being selected
when a project is loaded.

AddPendingEvent() needs to be used instead of the usual
ProcessEvent() in order to allow all handlers to finish processing
the Load Project event.

This is related to #686 which would still fail if there was nothing
selected.