Improve access to NodeSharePtr
When used in a for/range loop, this should be a const reference.

Passing the raw pointer is still the most common, but this does
add a few additional functions where you can pass in the shared
pointer and it will either process it directly, or get the raw pointer
and pass that to the function expecting it. This is for futrue -- it
means we have the option of changing to proccessing directly
using the shared pointer instead of indirectly working on the
raw pointer.

The other changes are to cleanup the use of for/ranged loops.