Move initialization code into OnInit
A wxUE-generated dialog can have the controls initialized in either the
ctor or the Create() function, so additional initialization should always be
in the OnInit() so that the dialog can be created as either a one-step or
two-step process.

The other change is to add enough rows for all fields the property already
has.