Add "no" as option for wxFrame center property
An empty field means to use the default which is wxBOTH, so to specify
a non-centered wxFrame, "no" must be used (wxDialog already does this).

Fix bug in handling XRC centered property

It was incorrectly returning from the function when the property was
encountered, but that would skip any subsequent properties.