Refactor tables and unknown prop handling
Tables are now at the top of the file where they can be easily found. Some
additions have been made to the property conversion as additional
XRC files are debugged and fixed.

Unknown properties are now processed in a separate function to make it
easier to find and work on.