Refactor ProcessUnknownProperty for clarity
Trying to find an option in a bunch of switch case statements is a lot
easier than searching through a bunch of string comparison statements.
Technically, this is slower, but it not even remotely noticeable, and being
able to find a specific option more than makes up for the additional
processing overhead.