Add debug command to duplicate form via XRC
This adds an additional way to verify that XRC export/import works as
expected. It take the currently selected form, converts it it to an XML
document using the XRC exporter, then imports that document into
a newly created node, and on success, inserts the node into the project
just below the previously selected form. To the tester, it will look similar
to what would happen if the form was duplicated.