Move internal cmake settings to their own file
Since the internal settings are really only useful for the maintainers,
having them inlined into the main CMakeLists.txt file just makes it harder
to read. Moving them out significantly reduces the clutter.

I also made a couple of fixes to support out-of-project builds by
adding a ${CMAKE_CURRENT_LIST_DIR}/ to some directories. Note that
these are *not* in the internal.cmake file, since that will always be done
as an in-project build.