Initial commit to support new folder generator
gen_folder and gen_sub_folder can be used as a parent for forms.

This commit has just enough code to verify that the new folder
can be displayed in the Navigation pane along with it's properties
in the Properties panel. It still needs a *LOT* of work before it
can be used without crashing.