Initially collapse all forms and folders on Project load
Forms within folders were initially all expanded -- this sets the initial state
of all forms, no matter what their parent is, to collapsed.

Folders are collapsed, but sub_folders are expanded (with the forms under
them collapsed). I'm expecting a folder to have few if any sub_folders, so
this should be a reasonable approach.