Add folder tool to Forms Ribbon Panel
This changes the "Images" collection in the Forms panel to "Other" and adds
the gen_folder command to that collection. This provides a way for a user
to create a new folder/sub_folder without needing to use the right-click
menu. If a form is currently selected, the folder becomes it's parent. If a
folder or sub_folder is selected, the new folder becomes the first child.