Enable generating multiple language types
The dialog box now uses checkboxes instead of radio buttons. The outer
function that calls the generators now handles all of the results. The function
declarations are all moved into a smaller header file, and the mainframe
functions changed to stand-alone (they had no reason to be part of
mainframe).