Refactor old/new XRC files for accuracy
Previously we parsed the old XRC file rather than doing a binary comparison
like the other generators do. For reasons unknown, parsing the file does
not generate the same string as writing the file. This new method is the
same as the other code generators, and has the advantage of not taking
the time to parse the original XRC file.