Enable code comparison for Python and Lua
This adds radio buttons to the internal dialog that compares code
using WinMerge. The buttons allow you to choose between C++,
Python and Lua for code generation comparison. This will make it
easier to check that a code generation change is creating the
desired change as well as checking that it does not cause
unintended code changes.