Add additional python properties
The wxPython version won't be used for now, but may be useful in the
future if additional wxWidgets functionality is supported in later versions
of wxPython.

The prefer_python_code can be used to minimize C++ categories, and
possibly to change what UI is presented to the user (such as the Event
handler dialog which accepts lambdas that can't be used in wxPython).