Add tab to C++ conditional code
This got broken when we shut off auto-indent for this function. Ultimately,
I need to see if we can *not* shut off the auto-indent code -- messing up
indents in python code will prevent the code from working, so much more
care needs to be taken with how indenting is handled.