Removed Import Tests
This was part of wxUiTesting, but it's a bit complicated to setup, and
currently it results in duplicate artwork since it is a separate project. While
I think it is important to implement at some point, I think we need a better
way to handle imports when they will be used for testing.