Initial rework of testing functionality
New MainTestDialog uses a Notebook with multiple pages to test
different controls. Outside of the dialog is list of events -- each control
should have at least one event handler which appends the name of
the event to this list. This ensures that wxPython has the write event
prefix (some need a library name in addition to the "wx." prefix.