Add 2-step  C++ wxFrame construction
Related to #892

This is now the same as wxDialog forms -- the dev can either construct an
empty class and then call Create(), or they can just construct the dialog
with any parameters they want to specify.

Note that this only applies to C++ code generation -- Python is 1-step for
all forms.
C