Change build name to wx_vadz
wx_utf8 has been merged into wxWidgets 3.3, so we only need the vadz
build if we're testing out a potential PR before it gets added to the main
wxWidgets dev branch. It could also be used for effectively and wxWidgets
build provided the setup.h and wxWidgets.cmake files match the
required build.