Try window_style if form has no style property
Not all forms have a style property, so if none is available, then check for a window_style, and if that is available, use it instead.

Closes #1019

Bug was encountered for wxPanel. That's currently the only form that doesn't have a style but does have a window_style, but in case a form is added in the future, the fix in this commit will handle it the same way as wxPanel.