Multiple changes to Undo actions
InsertNodeAction::Change() needs special processing for inserting an
embedded_image if the parent gen_Images is sorted.

Rewrite of SortImagesAction::Revert() to avoid deleting the gen_Images node
since all we really need to do is restore it's children.

AutoImagesAction::Revert() -- this is an experiment to see if we really need
special calls to the Navigational Panel.