Make the sorted option first
There's a special case where checking auto_update and sorted then undoing
the changes and redoing the changes results in an orphaned gen_images node. Changing the sort order makes it less likely for the problem to occur, so that we can finish doing other work on this branch before tracking down
the logic bug (which may not exist once other work is done).