Add second ctors that take a NodeSharedPtr
I keep having to get the raw pointer from a shared pointer to pass into a
ctor which immediately turns around and converts it back into a shared pointer. Instead, I created some Init() functions with two ctors so the Undo
Action could be created either way.

I didn't add it to all of them, but did to the ones being used a lot in this
branch.