Add comment to pch.h about which files to update
The information was already in docs/building notes but since I forgot to look
at that I can hardly expect anyone else to do so. Adding the comment in pch.h
where someone is most likely to make the change increases the chance
that all 4 files get changed at once.