Refactor as_wxArrayString()
Replace WX_PG_TOKENIZER2_BEGIN with actual code so someone
reading the code can know what is actually being done rather than
having to look up the macro to find out.

Provide a complete replacement using tt_string_view, and conditionalize so
that either version can be compiled.

For both versions if the string does not begin with a quote, assume it is
semi-colon separated.