Replace call to ConvertToArrayString
Conversion of old style to new style doesn't take very much code, so it
makes sense to inline it instead of calling a function (which was not as
efficient as the NodeProp version). If we convert other properties to use
semi-colons, then this new inserted code can become a lambda.