Don't add '_' before a number suffix
It's a bit odd to have a suffix like 'box_sizer_2' instead of  'box_sizer2'.
RuboCop goes so far as to say it's not part of snake_case naming
conventions. I've commented out the old code, and no longer add
the '_' before a numeric suffix. It works either way, so not a problem
to change back should a good reason for it be found.