Switch to frozen sets and maps
This is a test for the new frozen library -- these sets and maps are
created at compile time, so the don't any overhead to our app startup time.

I also refactored the code to take advantage of the new sets that replaced
the arrays which makes it possible to be more efficient in finding matching
styles.