Mark project as updated if check changed it
In Internal builds, RecursiveNodeCheck() is called and it could modify
the project file. In this commit, I moved the check functions into
the ProjectHandler, and then set m_isProject_updated to true if a change
was made.

I also added an INTERNAL_TESTING msg if row setting was removed from
wxGridSizer or wxFlexGridSizer.