Add the ability to hide a property
The ability to hide an inherited property already existed, this ties into that
by allowing the ability to hide an individual, non-inherited property. This can
be useful for internal properties. It can also help when a property has changed
since this allows the original property to be hidden so that it doesn't break
backwards compatibility to older projects.