Use wxNullBitmap instead of BitmapBundle()
The problem with using BitmapBundle() is it requires including bmpbndl.h just
to get the BitmapBundle() declaration even though it might not be used
anywhere else in the generated code. That requires conditionals around both
the header file and the Create() function. Using wxNullBitmap works fine
as an alternative to BitmapBundle() as far as wxWidgets 3.2 and beyond is
concerned.