Image List overhaul
This removes all extern statements from the header file, replacing them with
image_...() functions so that you no longer need to specify the size of the
data array to get a wxImage.

The bundle_ and image_ functions are now generated in the source file
after the data arrays since there are no more extern statements in the
header file.