Fix calculation on when an image function is needed
If there is an an Image List file, and an image can be accessed via bundle_...() then the m_Need function for that image is *not* set. This is to ensure that
only a class that is using an image not in the Image List will require a function
to load the image.