Don't add SVG to src if bundle is in Images List
This was affecting SetIcon for forms. The wxueBundleSVG() function was
being generated, but it wasn't being called because because a bundle
function in an Images List file was loading it. This doesn't matter with
C++17, but an earlier compiler produces a static function, and the compiler
complains about an unused static function.