Always record messages
The vector that stores messages is now global, and when g_pMsgLogging is
created, it will show the message log window if there were any messages.

This was done because previously we were getting warnings and errors but
never saw them unless a breakpoint was set when the message was added
and startup was run under a debugger.

Note that if the Splash window is open, you can't interact with the messages
log window.