Replace INTERNAL_TESTING with cmd-switch
This commit changes most of the code that was conditionalized using
INTERNAL_TESTING with a isTestingMenuEnabled() check at runtime.
This reduces the chance of broken builds on github when
INTERNAL_TESTING isn't used, as well as providing a lot more
functionality for testing with a release build. The downside is the
extra code and strings that are rarely used except by wxUiEditor
devs.

Note that INTERNAL_TESTING is still required to get ASSERTS in a
release build.