Drop PHP and replace it with Rust
Currently, PHP is build on wxWidgets 3.0 using PHP 5.4 -- making it
quite a bit behind wxWidgets 3.2 and PHP 7.  While the last Rust
port is not complete, it is at least built on 3.2, and in theory it should
be possible to extend it.

While I'm still not certain I'll support Rust, it's far more likely that
I would get that working than PHP.