Add check for testing switch
The Testing Menu is enabled if CMake has INTERNAL_TESTING
turned on (usually the case). The test switch is turned on if
test_menu is set on the command line.

Conditionalizing on isTestingSwitch() makes it possible to turn it
off while debugging to see how a Daily Release build would
behave.