Refactor Dialog positioning code for non-C++
Only C++ code generation exposes the dialog's Create() method, and
therefore it is currently the only language that needs to generate code
to handle the situation where Create() is called with a different
position or size then was specified for the default ctor.

Perl, Python, and Ruby code only provide the dialog's ctro, so
SetSizerAndFit() is all that's needed.