Move version/copyright into generated header file
Now that version.h is being generated by CMake, we don't want to
cause the entire project to be rebuilt because the version number
changed in pch.h. Instead, include the header file where it is actually
used.

txtAppname is now inline so that only one copy will be placed into
the executable.