Try to resolve error from gcc-14 building wxWidgets
Ultimately this will need to be fixed in wxWidgets, and will almost
certainly break again as soon as we switch to ▒FetchContent_Declare.

Daily builds don't break because they currently use the older gcc that
ships with Ubuntu 24.04, but the next Ubuntu release will undoubtedly
ship with a newer gcc that will generate this error.