Fix problems with C++ namespaces (#1654)
* Fix/Improve comments for GenHdrNameSpace()

Remove the incorrect // BUGBUG comment (tt_string_vector _is_ required!), and update
other comments to better explain this code section.

* Refactor GenerateClass() for clarity

This breaks out the code within GenerateClass() that creates all the #include statements
for both source and header file. Having this in a function makes it easier to maintain the
relevant code instead of trying to read through a long function.

* Fix formatting of CustomCtrl with a namespace

This pulls the namespace string out of the hdr_includes set so that it can be parsed
and formatted separately from the header files. It also ensures that the namespace
is generated after *all* of the #include files along with any preamble.

* Don't nest wxue_img namespace in class namespace

If the wxue_img namespace gets nested under a project/folder/form namespace,
then the wxue_img::bundle_function() for images will be invalid.