Fix/enhance external and embedded wxWidgets documentation (#1661)
* Change wxWidgets URL based on wxWidgets version

Use the C++ wxWidgets version to display either 3.2.8 or latest wxWidgets
documentation.

* Use latest, not trunk for wxWidgets docs

* Display C++ Doc URL in default browser

wxWidgets documentation now uses a captcha, and the IE embedded browser we
use will not display it.

* Use C++ for wxPerl documentation

wxPerl doesn't have it's own documentation web site the way wxPython and wxRuby
do, so we use the C++ documentation instead. Note that this displays the
documentation in the default browser, instead of embedding it.