Adjusting children of a wxGridBagSizer (#1666)
* Refactor to use class members, not local

Using m_node and m_parent means that when debugging you can set a watch
on them and have it valid for the entire class. The two members are set in the
constructor.

* Fix overlapping event ids

Like the comment in pch.h says, *ALL* event ids need to be set here or the update_ui event
can easily enable/disable the wrong item.

* Change gridbagsizer undo strings when moving

Instead of "Change" we specifically say Increase/Decrease