Split NewMdiForm into abstract/derived classes
I'm expecting the derived class to expand quite a bit, so it will be easier to work with
using an abstract/derived class pair instead of a single generated/edited class.

Note that we *can* still add derived class functions to the generated header file
that can be used for Binding event handlers.