Switch to m_cmdline_log for code generation
When generating code from the command line, UI cannot be displayed. This commit
moves the vector of warning/error messages into App::m_cmdline_log so that
code throughout wxUiEditor can check if is_Generating() is true, and if so, add
warning/messages to the vector which will be written out to the log file.