Additional work on MDI App creation generator (#1673)
* Split NewMdiForm into abstract/derived classes

I'm expecting the derived class to expand quite a bit, so it will be easier to work with
using an abstract/derived class pair instead of a single generated/edited class.

Note that we *can* still add derived class functions to the generated header file
that can be used for Binding event handlers.

* Minor refactoring

* Add 4 document and view generators

Image, RichText, SplitWindow, Scintilla

These are skeleton placeholders -- essentially a copy of TextCtrl just to get the files
and generators in place and ready for editing.

* Clarify comment for HeaderCode

* Update with more wxWidgets keywords

* Fix typo in type for document menubar

* Change hierarchy: app, doc, menu, and view classes

* Special-case DocViewApp class header creation

* Refactor

Mostly this is changing the hierarchy, but also name changes for code clarity.

* Change to unique text block names

* Add document and view templates

This also fixes some bugs in menu creation and function names.

* Fixed internal batch file for checking clang-format

The internal batch file I was using was missing a bunch of folders, thereby missing
formatting checks that weren't done before the code was committed.