Image handler refactoring (#1674)
* Update comments

* Add default destructor

* Remove support for animation .h files

This hasn't been supported in some time. This simply removes the (mostly
commented out) code that previously handled the format.

* Warn ig SVG file has parsing error when generating

Previously, this always displayed a message box which is invalid if generating code
from the command line. Now that error message gets written to the log file that is
written out after code generation is complete.

* Refactor using C++17 if-initializer

C++17 added an if statement with initializer enhancement that allows for more
readable way to check the results. Most of this commit changes to that format.

* Initial  ArtFolderChanged() function