Add new ttwx:: namespace functions/classes, refactor to avoid lint warnings (#1675)
* Allow tracking CMakeLists.txt files

* Add ttwx functions and classes

There are designed to replace most of the tt_ functions and classes while providing
support for the wxString class in utf8 mode.

* Replace tt::npos with std::string::npos

The latter name is longer, but is a standard (obviously).

* Allow implicit bool conversions

This allows std::is...() functions to be treated as a bool even though they return an int.

* Add the new ttwx files

* Remove owning memory warning

wxWIdgets takes ownership of a new allocated widget -- which clang-tidy doesn't
recognize. Since this happens a lot, remove the warning.

* Modernize NodeDeclaration class header

This also replaces the tt_string member with std::string -- that is what required the
various source files that needed to be modified as a result.

* Allow inconsistent enum assignments

Without this, there are a large number of problems listed for gen_enums.h which
uses assignments for overlow and other special cases.

* Refactor gen_enums

Use std::string_view instead of const char* for global std::maps. This increases memory
slightly, but potentially provides better performance since the string length is
determined at compile time.

* Modernize PropDeclaration header file

* Fix YAML syntax errors

.clang-tidy is a YAML file, so it needs to follow proper syntax or clang-tidy will fail.

* Refactor to use wxString and modern C++ style

* Add ttwx::StringVector class

Same as ViewVector, but based on std::vector<std::string> so that you can add, remove,
and edit the vector.

Changes to ViewVector and ttwx.h made to support two similar vector classes.

* Use std::string::npos, not tt::npos

* Initialize project to version 1.0

This always gets set when a project is loaded, but initializing it prevents lint warnings
about an uninitialized member variable.

* Add -cppcoreguidelines-owning-memory

wxWidgets will typically take ownership of newly created objects, so this warning
can be ignored. Unfortunately, there isn't a way to check it for non-wxWidgets
objects, though we normally use smart pointers for that.

* Options dialog needs to have std::string

Originally, generators were going to create options, but that's not currently
planned, so we need to make a copy from the XML sources.

* Fix value/help storage type

The original comment was wrong -- these strings are read in from the XML files,
which are discarded after initialization. As such, they must be copied -- using
std::string_view is invalid for a string buffer that gets deleted.

* Fix rare bug that showed up only while refactoring

The bug showed up when another bug temporarily caused a problem. While
it shouldn't be an issue in correct calls to tt::atoi, there is no point in not
tracking the potential empty string being converted to a digit.

* This fixes a clang complaint about a deprecated method

* Use anonymous namespace for static variables

* ttAssertionHandler is used in mainapp.cpp

It's a potential bug to declare a function in two different places since if the
definition changes, it won't be obvious that multiple declarations need to
be changed to match.

* wxString 3.3.x can be constructed from std::string_view

Replace .make_wxString with a temporary wxString constructor. That at least enables
the possibility of the allocation to be moved rather then copied to the final
wxString destination.

* Various other refactoring

* Fix formatting using clang-format

* Update clang to version 20

I'm using 21 on local machine, and it formats slightly differently then the version
18 that was previously in the github action.

* Fix incorrect DoozyX clang-version