Add options, refactor Project, add Catch2 test suite (#1677)
* Add minimum_compiler and trailing_return_type

These are for C++ code generation

* Expand C++ AI coding guidelines

Goal is that if AI is used to generate code, it should not result in clang-tidy warnings
about the generated code nor should a dev need to clean it up to match the
conventions of the rest of the codebase.

* Add ttwx::atoi() function

Because it accepts std::string_view and supports a leading 0x to indicate hexadecimal,
it's still a better alternative to std::stoi().

* Move all ttwx files into their own directory

These are still being updated, so it's easier if they are in their own directory.

* Use C++20 template/requires to avoid ambiguity

A "foo" string could either constructed as a std::string_view, or as a const wxString&.
Using the template with requires avoids that ambiguity.

* Add Catch2 testing for ttwx.cpp

This will build a ttwx_test executable that can be used to test ttwx functionality.
This initial commit just tests ttwx.cpp -- the two vector classes will be added next.

* Add more ttwx tests, configure when built

BUILD_TESTING must be set to ON before test_ttwx will be built. If turned on,
it always gets relinked -- so far I have been unable to find out why.

* Refactor for readability and modern coding style

* clang-format