Add hidden switches to call VerifyCodeGen()
This can be used to verify that code changes that should *NOT* cause any changes
to the code produced by code generation do not, in fact, change them. This is
different from when you expect changes, but if you expect changes in one language,
but not others, you can Verify no changes to the others.