Verify code generation (#1682)
* Place ".wxui" extension in a global std::stringview

This avoids typos and makes it clear what the extension is for.

* Add hidden switches to call VerifyCodeGen()

This can be used to verify that code changes that should *NOT* cause any changes
to the code produced by code generation do not, in fact, change them. This is
different from when you expect changes, but if you expect changes in one language,
but not others, you can Verify no changes to the others.

* Use a shell script to format C++ files

Most/all Agents can't access VSCode commands, so it's easier to just save the file
and run a shell script to format the file.

* Explain the files in the generate folder

This is just a general summary of what the files in this directory are for. We assume
a reader will be able to figure out what the ones in the writers/ folder do.

* Add functions and class needed by Verify

While VerifyCodeGen is the only one that needs this now, it's definitely possible that
another code section might need to use them.