Reduce Function complexity in ProjectHandler, refactor utils.cpp/utils.h (#1685)
* Refactor to reduce function complexity

Switch to private helper functions instead of using a large lambda or function.

* Add structures to avoid adjacent parameters

Since we require C++20, this allows us to use designated initializers to be sure the
correct Node* pointers are being passed to functions.

* Change co-pilot build instructions

Using VSCode and PowerShell as the terminal, copilot was not recognizing that a build
failed. These expanded instructions are designed to fix that, though we may need
to tweak the search for FAILURE indicators if errors still slip through.

* Don't warn about magic numbers

There are times when the reason for the number is obvious, and changing it to a
constexpr isn't very helpful. Other times, it's quite valuable. I'll leave it to the devs
to decide when they are coding whether to use magic numbers or not.