Refactor, and fix Ruby regression with missing end statement
Refactoring to reduce the complexity so that I could follow the logic for dealing with
various scenarios with/without the required end statement for the class.

While only Ruby needs to write the end statement outside of the final comment block,
the refactoring affects all code generation. As far as I can tell, it doesn't change the
previous functionality.