Fix lack of end statement in generated Ruby code. (#1688)
* Refactor FileCodeWriter::WriteFile

This function was too complex complicating trying to fix problems. I also replaced
all of the `tt_` code with std:: and wxWidgets classes. Changes also made to other
functions to replace tt_ classes with more standard ones.

* Conditionalize OnVerifyTTWX

Currently, this is for debug builds only.

* Refactor, and fix Ruby regression with missing end statement

Refactoring to reduce the complexity so that I could follow the logic for dealing with
various scenarios with/without the required end statement for the class.

While only Ruby needs to write the end statement outside of the final comment block,
the refactoring affects all code generation. As far as I can tell, it doesn't change the
previous functionality.