Possible workaround for odd Unix compiler problem
Unix runner failed stating that the this pointer is null calling m_frame->Close() near the end of the OnRun() function. I'm a bit mystified as to why the compiler thinks that since we never reach that code unless m_frame is valid. So, this is just a possible fix to test out.