Node class refactoring (#1694)
* Reduce complexity of some functions

New helper functions added to make it easier to maintain the code.

* Switch from tt_ classes to std:: and wxString classes

* Added comment about threading analysis

After reviewing, threading the startup XML processing isn't practical. In case I or someone
else wants to consider it again, I added the comments as to why it's not practical.

* Refactor for clarity, reduced complexity

* Remove requirement for [[nodiscard]]

It's sometimes useful to have to prevent coding errors, but the blanket requirement
even for just functions that return true is more annoying than useful.

* Refactor for readability/maintainability

* Replace some tt_string functions in Node class