Update github Runners (#1698)
* Add ccache --show-config to try to see why it's failing

The second run reported only a fraction of the possible hits, and all uncallable caching. This commit will try to fix that...

* Move and edit ccache setup

Probably overkill on the sloppiness setting, so reduced to a smaller set.

Still trying to get it to cache wxWidgets object files.

* Turned on had-link

* Copy changes from ubuntu action

* Change the sloppiness setting

It applies to all platforms, but should use fewer conditions then previously set. The sole
reason is to get precompiled headers to work.

* Convert the action to also build ubuntu version

With the new ccache setttings, it fast enough to be useful.