Improve AdjustMemberNameForLanguage()
After the refactor, new nodes were not longer be created with a var_name setting.
This commit changes the way the function works to handle a new node that
doesn't have var_name set yet, and to fix it up if needed based on preferred
generator language and then check for duplicate names.