Node Creation Refactoring and Code Quality Improvements (#1703)
* Moved NodesParentChild struct to nodes.h

Since this is a Nodes parameter function, it doesn't make any sense elsewhere

* Increase MSVC Linker stack size

Linker is sometimes running out of memory. Increasing it to 8MB should resolve the
issue, and that's the standard stack size anyway for Unix and Mac.

* Change to scoped enum

The C++20 "using enum Node::Validity;" makes scoped enums much less wordy
while still providing the needed type safety.

* Switch to std::stringview

* Refactor for brevity, add performance section

Goal is to prefer brevity over human readability to minimize how much copilot needs
to parse. Performance section added because anything in the nodes/ and generate/
are used in generating code. Performance in these code sections is critical to keep
generating time for large projects (like ours) from taking a long time.

* Improve AdjustMemberNameForLanguage()

After the refactor, new nodes were not longer be created with a var_name setting.
This commit changes the way the function works to handle a new node that
doesn't have var_name set yet, and to fix it up if needed based on preferred
generator language and then check for duplicate names.