Enhanced Code Verification with Detailed Diff Output (#1708)
* Use new diff code in Verify function

This is primarily for use with AI where we want to confirm that refactoring did not
break any existing code generation functionality. It can also be used to quickly
tell if a fix to code generation fixed the problem as we expected.

* Make certain AI doesn't use tt class functions in new code

We're moving away from tt entirely so that code is easier to understand and more
reusable.

* Add instructions for using verify command

The is to verify that refactoring by AI doesn't break code generation.