Update Perl information
The updated Supported file has 'y' and 'n' in the Perl column indication where it is available or not. The 'y' means wxUiEditor should be able to generate code for it.

The agents update gives more information about how an agent can help with updating the code to generate wxPerl code.