Add more functions to ttwx namespace (#1726)
* Fix file corruption

Unfortunately, the last time I had an issue with the way the architect was handling the plan, I asked it to update it's instructions, and it corrupted the file. This is the second time an agent has corrupted a file -- need to check that before commiting.

* Add line number accuracy, remove file links

Unfortunately, file links are completely unreliable, so I removed them. Line number accuracy gets messed up when the file is compressed in memory, or is only partial. Using read_file helps, grep adds the the accuracy. Both will hopefully make the line numbers accurate.

* Fix invalid warning about unreachable line

FAIL_MSG does *not* terminate the app (unless the user asks it to).

* Update ttwx with conversion functions

These are the functions identified as being need to help with the transition from tt_ classes to more standard classes, with the ttwx namespace functions occasionally providing some additional functions.

* Fix clang-format

This shouldn't be necessary since I always format on save, but somehow it's slipping through... :sad: