Refactor code generation architecture to unify display, comparison, and file writing through GenResults orchestration class (#1730)
* Initial refactoring to fix comparison logic

Comparison now uses the recently refactored FileCodeWriter class

* Add a set of all valid GEN_LANG languages

By using gen_lang_set() to verify that the language is supported, it becomes possible to add a language without having to hunt for every file that processes a GEN_LANG language.

* Turn GenResults into a class

I'll be adding more methods directly to this class instead of having them spread around in other source files. As such, it makes more sense to convert it to a class, along with using proper private members with access methods instead of callers accessing members directly.

* Remove all check_build settings

The github Unix build now uses clang, so we are already getting a non-MSVC compiler with good error reporting. It's also what's run during a PR, so we don't need to run this locally any more.

* Refactor GenResults class

With this extensive commit, display, comparison, and file writing all go through the same code path.

* Refactor to modernizie, replace tt_ classes/functions

* Cast tt_string to std::string

std::format won't accept a tt_string as an argument

* Fix formatting

* Cast tt_string to std::string for std::format

* Fix another tt_string that std::format can't handle

Note that these fixes are for gcc on Unix -- MSVC works just fine without having to cast.

* Fix another std::format/tt_string issue

* Add missing parenthesis

* Cast, not convert tt_string to std::string

* Use tmp string to extract std:::string from tuple

* Make fixes for gcc on Unix