Modernize, replace most tt and ttwx funcs/classes with wxue_namespace versions (#1734)
* Complete batch 1-4 of modernization pass

* Complete modernization pass

* Add more functions to ttwx

* Create initial design decision docs

* Add 3 functions from tt_view_vector

These missing functions were preventing conversion

* Process all files to utilize new wxue_namespace

* Fix/reduce compiler warnings

* Minor changes

* Fix warnings and errors on Ubuntu

* Switch to clang-20

MSVC and gcc 15.1.1 work fine with the current codebase, but clang is reporting problems with std::format. This commit updates it to version 20 to see if that will resolve the issue.

* Fix broken github action

* Various fixes

This include fixes that should work with clang-20 and it's strict std::format() requirements.

* Fix stricter type error in clang-20